# src/main/contracts/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Delivery
  description: Отвечает за учет курьеров, деспетчеризацию доставок, доставку
paths:
...
  /api/v1/couriers:
    post:
      summary: Добавить курьера
      description: Позволяет добавить курьера
      operationId: CreateCourier
      tags:
        - CreateCourier
      requestBody:
        description: Курьер
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCourier'
      responses:
        '201':
          description: Курьер успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCourierResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Конфликт при создании курьера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'       
    get:
      summary: Получить всех курьеров
      description: Позволяет получить всех курьеров
      operationId: GetCouriers
      tags:
        - GetCouriers
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Courier'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Location:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
          description: X
          minimum: 0  # Валидация минимального значения для x
        y:
          type: integer
          description: Y
          minimum: 0  # Валидация минимального значения для y
...
    NewCourier:
      type: object
      required:
        - name
        - speed
      properties:
        name:
          type: string
          description: Имя
          minLength: 1  # Валидация на минимальную длину
        speed:
          type: integer
          description: Скорость
          minimum: 1
    Courier:
      type: object
      required:
        - id
        - name
        - location
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор
        name:
          type: string
          description: Имя
        location:
          $ref: '#/components/schemas/Location'
          description: Геолокация
...
    CreateCourierResponse:
      type: object
      properties:
        courierId:
          type: string
          format: uuid                
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Код ошибки
        message:
          type: string
          description: Текст ошибки